+++ pkg-py/src/shinychat/playwright/_chat.py
@@ -95,7 +95,9 @@ class Chat(UiBase):
         self.loc_latest_message = self.loc_messages.locator("> :last-child")
         self.loc_input_container = self.loc.locator("> shiny-chat-input")
         self.loc_input = self.loc_input_container.locator("textarea")
-        self.loc_input_button = self.loc_input_container.locator("button")
+        self.loc_input_button = self.loc_input_container.locator(
+            ".shiny-chat-btn-send"
+        )
 
     def expect_latest_message(
         self,